backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
scapy==2.5.0
requests==2.31.0
dnspython==2.4.2
pydantic==2.3.0
cryptography==41.0.4
psutil==5.9.5
